+++ demo2/webroot/frontend/src/Components/ComponentA/ComponentA.js
@@ -1,0 +1,10 @@
+/**
+ * Created by Martin Neundorfer on 10.09.2018.
+ * For LABOR.digital
+ */
+document.addEventListener('DOMContentLoaded', () => {
+	console.log('loaded');
+	document.querySelector('.componentA').addEventListener('click', e => {
+		alert('component A: ' + e.type);
+	});
+}, false);