+++ demo2/webroot/frontend/src/DynamicComponents/ComponentC/ComponentC.js
@@ -1,0 +1,5 @@
+/**
+ * Created by Martin Neundorfer on 10.09.2018.
+ * For LABOR.digital
+ */
+alert('I\'m loaded via ajax! :D');