+++ demo2/webroot/frontend/src/app.js
@@ -1,0 +1,9 @@
+/**
+ * Created by Martin Neundorfer on 10.09.2018.
+ * For LABOR.digital
+ */
+// Import globals
+import './Globals/Globals.sass';
+
+// Import all components
+import './Components/*/*.js'